libc: minimal: move strtok_r() to posix
The strtok_r() function is POSIX. It's not part of ISO C.

This fixes an issue where the minimal libc would not compile
without defining `_POSIX_C_SOURCE=200809` in an ad-hoc
manner.
