+++ lib/libc/minimal/source/string/string.c
@@ -151,45 +151,6 @@ int strncmp(const char *s1, const char *s2, size_t n)
 	return (n == 0) ? 0 : (*s1 - *s2);
 }
 
-/**
- * @brief Separate `str` by any char in `sep` and return NULL terminated
- * sections. Consecutive `sep` chars in `str` are treated as a single
- * separator.
- *
- * @return pointer to NULL terminated string or NULL on errors.
- */
-char *strtok_r(char *str, const char *sep, char **state)
-{
-	char *start, *end;
-
-	start = str ? str : *state;
-
-	/* skip leading delimiters */
-	while (*start && strchr(sep, *start)) {
-		start++;
-	}
-
-	if (*start == '\0') {
-		*state = start;
-		return NULL;
-	}
-
-	/* look for token chars */
-	end = start;
-	while (*end && !strchr(sep, *end)) {
-		end++;
-	}
-
-	if (*end != '\0') {
-		*end = '\0';
-		*state = end + 1;
-	} else {
-		*state = end;
-	}
-
-	return start;
-}
-
 char *strcat(char *ZRESTRICT dest, const char *ZRESTRICT src)
 {
 	strcpy(dest + strlen(dest), src);
