+++ tests/posix/cx/prj.conf
@@ -1,0 +1,2 @@
+CONFIG_ZTEST=y
+CONFIG_POSIX_CX_STRING_STRTOK_R=y
